first draft restructure
