+++ src/actuators/Buzzer.java
@@ -1,0 +1,5 @@
+package actuators;
+
+public class Buzzer {
+
+}
