+++ src/actuators/Motor.java
@@ -1,0 +1,10 @@
+package actuators;
+
+import TI.BoeBot;
+
+public class Motor extends TI.Servo{
+
+    public Motor(int pin){
+        super(pin);
+    }
+}
