+++ src/actuators/NeoPixel.java
@@ -1,0 +1,7 @@
+package actuators;
+
+public class NeoPixel {
+    public NeoPixel(){
+
+    }
+}
