+++ src/head/Controller.java
@@ -1,13 +1,15 @@
+package head;
+
 import TI.BoeBot;
 import TI.PinMode;
 
-public class Engine implements Updateable{
+public class Controller implements Updateable {
 
     public Boolean isRunning;
     private Zoomer zoomer;
 
     public EmergencyStop emergencyStop;
-    Engine() {
+    public Controller() {
         BoeBot.setMode(1, PinMode.Input);
         this.isRunning = true;
         this.zoomer = new Zoomer(10, 11);
@@ -43,8 +45,8 @@ public class Engine implements Updateable{
 
 
 //    public static void main(String[] args) {
-//        Zoomer zoomer = new Zoomer(12, 14);
-//        EmergencyStop emergencyStop = new EmergencyStop(0);
+//        head.Zoomer zoomer = new head.Zoomer(12, 14);
+//        head.EmergencyStop emergencyStop = new head.EmergencyStop(0);
 //
 //        while (true) {
 //
@@ -52,7 +54,7 @@ public class Engine implements Updateable{
 //            zoomer.update(14);
 //            if(emergencyStop.check()){
 //                //check for emergency stop press, stop the wheels and break the loop
-//                MotorAansturen.stop();
+//                head.MotorAansturen.stop();
 //                break;
 //            }
 //            BoeBot.wait(1);
