+++ src/head/Updateable.java
@@ -1,3 +1,5 @@
+package head;
+
 public interface Updateable {
     void update();
 }
