+++ src/sensors/Button.java
@@ -1,0 +1,11 @@
+package sensors;
+
+import TI.BoeBot;
+import TI.PinMode;
+
+public class Button {
+
+    public Button(int pin){
+        BoeBot.setMode(pin, PinMode.Input);
+    }
+}
