+++ src/sensors/LineDetector.java
@@ -1,0 +1,4 @@
+package sensors;
+
+public class LineDetector {
+}
